fix: fix the priority stops ch101 issue
When the task priorities for VL53L1X and CH101 are set to the same
level, the sampling task for CH101 gets interrupted. It is suspected
that one or more `DATA_READY_BIT` flags are not being set for some
reason. Temporarily, this issue can be mitigated by increasing the
priority of the CH101 task. However, there is room for optimization.

Signed-off-by: Dennis Liu <dennis48161025@gmail.com>